--- a/dev/clj_deconstruct_sigs/dev/data.clj
+++ b/dev/clj_deconstruct_sigs/dev/data.clj
@@ -2,7 +2,8 @@
   (:require [clj-http.client :as http]
             [clojure.data.csv :as csv]
             [clojure.string :as string]
-            [clojure.java.io :as io]))
+            [clojure.java.io :as io]
+            [fipp.clojure :as fipp]))
 
 (def cosmic-sigs-endpoint
   "https://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt")
@@ -40,3 +41,49 @@
     (spit "src/clj_deconstruct_sigs/data/generated.clj"
           (str (list 'ns 'clj-deconstruct-sigs.data.generated) "\n\n"
                (list 'def 'latest-cosmic-signatures (vec (map vec without-header)))))))
+
+
+(defn write-test-data-cljc!
+  "Converts CSV files containing test datasets into a cljc file."
+  ([ns']
+   (write-test-data-cljc! ns' {}))
+  ([ns' opts]
+   (let [test-cosmic-signatures (->> (io/resource "test-cosmic-signature.csv")
+                                     slurp
+                                     csv/read-csv
+                                     drop-headers
+                                     (mapv vec))
+         random-tumor-samples (->> (io/resource "random-tumor-samples.csv")
+                                   slurp
+                                   csv/read-csv
+                                   drop-headers
+                                   (mapv vec))
+         answers (mapv #(-> (str "answer/answer-" (inc ^long %) ".csv")
+                            io/resource
+                            slurp
+                            csv/read-csv
+                            drop-headers
+                            first
+                            vec)
+                       (range 100))
+         f (-> ns'
+               str
+               (string/replace #"\-" "_")
+               (string/replace #"\." "/")
+               (as-> s (str "test/" s ".cljc")))]
+     (with-open [w (io/writer f)]
+       (binding [*out* w]
+         (fipp/pprint `(~'ns ~ns'))
+         (newline)
+         (fipp/pprint `(def ~'test-cosmic-signatures ~test-cosmic-signatures) opts)
+         (newline)
+         (fipp/pprint `(def ~'random-tumor-samples ~random-tumor-samples) opts)
+         (newline)
+         (fipp/pprint `(def ~'answers ~answers) opts))))))
+
+
+(comment
+
+  (write-test-data-cljc! 'clj-deconstruct-sigs.data.cosmic-test {:width 5000})
+
+  )