--- a/project.clj
+++ b/project.clj
@@ -9,7 +9,12 @@
                    :all (constantly true)}
   :profiles {:dev {:source-paths ["dev"]
                    :resource-paths ["dev-resources"]
-                   :dependencies [[clj-http "3.9.1"]
-                                  [org.clojure/data.csv "0.1.4"]]}
+                   :dependencies [[org.clojure/clojure "1.10.0"]
+                                  [clj-http "3.9.1"]
+                                  [org.clojure/data.csv "0.1.4"]
+                                  [net.totakke/libra "0.1.1"]
+                                  [criterium "0.4.4"]]
+                   :plugins [[net.totakke/lein-libra "0.1.2"]]
+                   :libra {:bench-paths ["bench"]}}
              :1.9 {:dependencies [[org.clojure/clojure "1.9.0"]]}
              :1.10 {:dependencies [[org.clojure/clojure "1.10.0"]]}})