--- a/test/clj_deconstruct_sigs/core_test.clj
+++ b/test/clj_deconstruct_sigs/core_test.clj
@@ -1,13 +1,15 @@
 (ns clj-deconstruct-sigs.core-test
-  (:require [clj-deconstruct-sigs.core :refer :all]
-            [clojure.core.matrix :as m]
+  (:require [clojure.test :refer :all]
+            [clj-deconstruct-sigs.core :as core]
             [clojure.data.csv :as csv]
             [clojure.java.io :as io]
-            [clojure.test :refer :all]
             [clj-deconstruct-sigs.data.cosmic :as data-cosmic]
             [clj-deconstruct-sigs.dev.data :as dev-data]))
 
-(def test-cosmic-sigatures
+(def ^:const ^:private which-signatures-keyset
+  #{:seed-idx :weights :weights* :product :unknown :diff :error-sum})
+
+(def test-cosmic-signatures
   (-> (io/resource "test-cosmic-signature.csv")
       slurp
       csv/read-csv
@@ -20,38 +22,36 @@
       dev-data/drop-headers))
 
 (deftest basic-which-signatures-test
-  (let [result (which-signatures
+  (let [result (core/which-signatures
                 (first random-tumor-samples)
-                test-cosmic-sigatures)]
-    (is (= (set (keys result))
-           #{:seed-idx
-             :weights
-             :weights*
-             :product
-             :unknown
-             :diff
-             :error-sum}))))
+                test-cosmic-signatures)]
+    (is (= which-signatures-keyset (set (keys result))))))
 
 (deftest cosmic-signatures-test
-  (let [result (which-signatures
+  (let [result (core/which-signatures
                 (first random-tumor-samples)
                 data-cosmic/cosmic-signatures)]
-    (is (= (set (keys result))
-           #{:seed-idx
-             :weights
-             :weights*
-             :product
-             :unknown
-             :diff
-             :error-sum}))))
+    (is (= which-signatures-keyset (set (keys result))))))
 
-(defn to-vec [m]
-  (->>  m
-        (sort-by first)
-        (map second)))
+(defn- to-vec [m]
+  (->> m (sort-by first) (map second)))
 
-(defn measure-difference [v1 v2]
-  (Math/sqrt (m/esum (m/pow (m/sub v1 v2) 2))))
+(defn- measure-difference [v1 v2]
+  (Math/sqrt (reduce + 0.0 (map (comp #(Math/pow % 2.0) -) v1 v2))))
+
+(deftest compare-reference-result-once-test
+  (let [answer-weights (-> "answer/answer-1.csv"
+                           io/resource
+                           slurp
+                           csv/read-csv
+                           dev-data/drop-headers
+                           first)
+        result (-> random-tumor-samples
+                   first
+                   (core/which-signatures test-cosmic-signatures))]
+    (is (< (measure-difference answer-weights (to-vec (:weights result))) 1e-15))
+    (is (= 25 (:seed-idx result)))
+    (is (< (:error-sum result) 0.02369))))
 
 (deftest ^:slow compare-reference-result-test
   (let [answers (map #(-> (io/resource (str "answer/answer-" (inc %) ".csv"))
@@ -60,15 +60,11 @@
                           dev-data/drop-headers
                           first)
                      (range 100))
-        my-answers (pmap #(to-vec (:weights (which-signatures % test-cosmic-sigatures)))
+        my-answers (pmap #(to-vec (:weights (core/which-signatures % test-cosmic-sigatures)))
                          random-tumor-samples)]
-    ;; For debugging as the computation in clojure takes a long time...
-    (with-open [w (io/writer "/tmp/clj-output.csv")]
-      (csv/write-csv w my-answers))
-    (is (< (m/esum (map #(measure-difference %1 %2) answers my-answers))
-           0.00001))))
+    (is (< (reduce + (map #(measure-difference %1 %2) answers my-answers)) 1e-5))))
 
 (deftest signature->vector-test
-  (is (= (signature->vector {}) nil))
-  (is (= (signature->vector {{:ref \C :alt \G :before \T :after \A} 3})
-         '(0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 1.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0 0.0))))
+  (is (= (core/signature->vector {}) nil))
+  (is (= (core/signature->vector {{:ref \C :alt \G :before \T :after \A} 3})
+         (-> (repeat 96 0.0) vec (assoc 28 1.0)))))