--- a/test/clj_deconstruct_sigs/test_runner.cljs
+++ b/test/clj_deconstruct_sigs/test_runner.cljs
@@ -1,0 +1,6 @@
+(ns clj-deconstruct-sigs.test-runner
+  (:require  [doo.runner :refer-macros [doo-tests]]
+             [clj-deconstruct-sigs.core-test]))
+
+(doo-tests 'clj-deconstruct-sigs.core-test)
+