… .travis.yml
… README.md
… src/clj_deconstruct_sigs/data/cosmic.clj
… src/clj_deconstruct_sigs/data/cosmic.cljc
… test/clj_deconstruct_sigs/core_test.cljc
… test/clj_deconstruct_sigs/data/cosmic_test.cljc