This PR improves performance of `which-signatures` by 100x.

The biggest bottleneck was calculating errors `|Aw - x|^2`.
`A` is relatively low-dimentional (96x25) and the overhead of type conversion / cloning was too large.
I fixed this by
- using `double-array` and primitive arithmetic operations instead of `core.matrix`
- refactoring some fns to reduce intermediate objects

This PR also includes
- benchmarking scripts
- cljs support


#### `lein libra`

##### master

```
which-signatures-bench-1
  time: 16.204357 sec, sd: 425.217328 µs
which-signatures-bench-100
  time: 2349.242194 sec, sd: 23.598227 sec
```

##### PR

```
which-signatures-bench-1
  time: 160.306164 ms, sd: 2.695163 µs
which-signatures-bench-100
  time: 22.940454 sec, sd: 6.505755 ms
```