@iku000888 2019-02-15
Thanks! It also crossed my mind that `Performance and implementation` section in the readme should be updated :sunglasses: 

Otherwise than that I am going to leave it to @federkasten for the timing etc to merge and release this.