@alumi 2019-02-19
I've deleted the section. @federkasten Please review it when you have time 🙏